+++ src/pages/activity-detail/index.tsx
@@ -1,0 +1,5 @@
+function ActivityDetail() {
+  return <div>체험 상세 페이지</div>;
+}
+
+export default ActivityDetail;
